common: drop config in call_stacks_analysis
